requirements.txt
openai>=1.0.0
qdrant-client>=1.7.0
tiktoken>=0.5.0
python-dotenv>=1.0.0
python-docx>=1.0.0
PyPDF2>=3.0.0
pyyaml>=6.0
fastapi>=0.100.0
uvicorn>=0.23.0
mcp[cli]>=1.0.0
starlette>=0.27.0
cohere>=5.0.0
